fixed a few compiler warnings with Visual Studio.

+++ libxvcalc/XvcOperatorCall.c
@@ -175,7 +175,7 @@ static XvcNumber Dice(XvcNumber left, XvcNumber right, jmp_buf jb)
 	/* Due to the grammar, right is guarenteed to be an integer. */
 	faces = right.i;
 
-	if (!has_seeded) srand(time(NULL));
+	if (!has_seeded) srand((unsigned int)time(NULL));
 	for (i = 1; i <= count; i++) running += (rand() % faces) + 1;
 	rVal.type = 'i';
 	rVal.i = running;
